reports/pyproject.toml
[tool.poetry]
name = "reports"
version = "0.1.0"
description = ""
authors = ["Edwar Martinez Vale <edwar_kenedy@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^1.4.46"
openpyxl = "^3.0.10"
python-socketio = "^5.7.2"
aiohttp = "^3.8.3"

[tool.pyright]
venvPath = "./"
venv = ".venv"
reportOptionalCall = 'none'
executionEnvironments = [
  { root = "reports" }
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
